+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added `--data-dirs` flag to the varlogsn to indicates the required data path to the storage node. (#VARLOG-769/#694)
+- Added `--volume-strict-check` flag to the varlogsn to check the volumes strictly. (#VARLOG-769/#694)
+
 ### Changed
 - Operation script `start_varlogsn.py` will not run the storage node if there are any missing data directories. (#VARLOG-765/#693)